completed

+++ Spring Fundamentals-sept 2021/Exam Prep/Spring Fundamentals Exam - 21 February 2021/examAndery01/src/main/java/com/exam/examandery01/web/ItemController.java
@@ -1,12 +1,12 @@
 package com.exam.examandery01.web;
 
-
 import com.exam.examandery01.model.binding.ItemAddBindingModel;
 import com.exam.examandery01.model.services.ItemServiceModel;
 import com.exam.examandery01.model.view.ItemViewModel;
 import com.exam.examandery01.service.ItemService;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
@@ -14,7 +14,6 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
-import java.util.List;
 
 @Controller
 @RequestMapping("/items")
@@ -29,13 +28,19 @@ public class ItemController {
     }
 
     @GetMapping("/add")
-    public String add() {
+    public String add(Model model, HttpSession session) {
+        if (session.getAttribute("user") == null) {
+            return "redirect:/users/login";
+        }
 
+        if (!model.containsAttribute("itemAddBindingModel")){
+            model.addAttribute("itemAddBindingModel",new ItemAddBindingModel());
+        }
         return "add-item";
     }
 
     @PostMapping("/add")
-    public String addConfirm(@Valid @ModelAttribute("model") ItemAddBindingModel model,
+    public String addConfirm(@Valid @ModelAttribute("itemAddBindingModel") ItemAddBindingModel itemAddBindingModel,
                              BindingResult bindingResult, RedirectAttributes redirectAttributes,
                              HttpSession session) {
         if (session.getAttribute("user") == null) {
@@ -43,16 +48,21 @@ public class ItemController {
         }
 
         if (bindingResult.hasErrors()) {
+            redirectAttributes.addFlashAttribute("itemAddBindingModel",itemAddBindingModel);
+            redirectAttributes.addFlashAttribute("org.springframework.validation.BindingResult.itemAddBindingModel", bindingResult);
+
             return "redirect:add";
         }
-        String id = this.itemService.add(modelMapper.map(model, ItemServiceModel.class));
+        String id = this.itemService.add(modelMapper.map(itemAddBindingModel, ItemServiceModel.class));
         return "redirect:/items/details/?id=" + id;
     }
 
     @GetMapping("/details")
-    public ModelAndView details(@RequestParam("id") String id, ModelAndView modelAndView) {
+    public ModelAndView details(@RequestParam("id") String id, ModelAndView modelAndView,HttpSession session) {
+        if (session.getAttribute("user") == null) {
+           modelAndView.setViewName("login");
+        }
         if (id == null) {
-
             modelAndView.setViewName("home");
         } else {
             ItemViewModel viewModel = this.itemService.GetItemById(id);
@@ -67,8 +77,10 @@ public class ItemController {
     }
 
     @GetMapping("/delete")
-    public String delete(@RequestParam("id") String id) {
+    public String delete(@RequestParam("id") String id, HttpSession session) {
+        if (session.getAttribute("user")!=null){
         this.itemService.delete(id);
+        }
         return "redirect:/";
     }
